use toggle panel from makepad
