+++ src/shared/mod.rs
@@ -8,7 +8,6 @@ pub mod modal;
 pub mod portal;
 pub mod resource_imports;
 pub mod styles;
-pub mod toggle_panel;
 pub mod tooltip;
 pub mod utils;
 pub mod widgets;
@@ -23,5 +22,4 @@ pub fn live_design(cx: &mut Cx) {
     download_notification_popup::live_design(cx);
     tooltip::live_design(cx);
     desktop_buttons::live_design(cx);
-    toggle_panel::live_design(cx);
 }
